Move loadDatasetsAndAnnotations to FetchData.js
